Fixed bug chrash on opening red door

+++ src/doom/r_plane.cpp
@@ -471,7 +471,7 @@ void R_DrawPlanes(void)
                 dc_yl = pl->top[x];
                 dc_yh = pl->bottom[x];
 
-                if (dc_yl <= dc_yh) // [crispy] 32-bit integer math
+                if (static_cast<unsigned>(dc_yl) <= dc_yh) // [crispy] 32-bit integer math
                 {
                     angle     = ((an + xtoviewangle[x]) ^ flip) >> ANGLETOSKYSHIFT;
                     dc_x      = x;
